industry-mail-system/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0
google-auth==2.26.0

# Added for scheduler and optional OpenAI client
apscheduler==3.10.1
# (openai removed; using Gemini via HTTP API key)
